requirements.txt
flask>=2.0.0
google-generativeai>=0.3.0
google-genai>=0.1.0
openrouter>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
